protect regression prints for parallel

+++ regress.cpp
@@ -1,13 +1,14 @@
 bool check_regression(std::string const& prefix, Mesh* mesh,
     Real tol, Real floor) {
+  auto comm = mesh->comm();
   auto goldpath = prefix + ".osh";
   if (!directory_exists(goldpath.c_str())) {
-    if (mesh->comm()->size() == 1) {
+    if (comm->size() == 1) {
       binary::write(goldpath, mesh);
       std::cout << "gold path \"" << goldpath << "\" did not exist yet,\n";
       std::cout << "created it from this run.\n";
       return true;
-    } else if (mesh->comm()->rank() == 0) {
+    } else if (comm->rank() == 0) {
       auto tmppath = prefix + "_tmp.osh";
       binary::write(tmppath, mesh);
       std::cout << "gold path \"" << goldpath;
@@ -18,25 +19,31 @@ bool check_regression(std::string const& prefix, Mesh* mesh,
     }
   }
   Mesh mesh2;
-  binary::read(goldpath, mesh->comm(), &mesh2);
+  binary::read(goldpath, comm, &mesh2);
   if (compare_meshes(mesh, &mesh2, tol, floor, false, true)) {
-    std::cout << "This run matches gold \"" << goldpath << "\"\n";
+    if (comm->rank() == 0) {
+      std::cout << "This run matches gold \"" << goldpath << "\"\n";
+    }
     return true;
   }
   std::cout << "meshes not exactly the same, trying superset...\n";
   if (compare_meshes(mesh, &mesh2, tol, floor, true, true)) {
     auto newpath = prefix + "_new.osh";
     binary::write(newpath, mesh);
-    std::cout << "This run, stored at \"" << newpath << "\",\n";
-    std::cout << "has more tags than \"" << goldpath << "\"\n";
-    std::cout << "It should probably be made the new gold, like this:\n";
-    std::cout << "  rm -rf \"" << goldpath << "\"\n",
-    std::cout << "  mv \"" << newpath << "\" \"" << goldpath << "\"\n";
+    if (comm->rank() == 0) {
+      std::cout << "This run, stored at \"" << newpath << "\",\n";
+      std::cout << "has more tags than \"" << goldpath << "\"\n";
+      std::cout << "It should probably be made the new gold, like this:\n";
+      std::cout << "  rm -rf \"" << goldpath << "\"\n",
+      std::cout << "  mv \"" << newpath << "\" \"" << goldpath << "\"\n";
+    }
     return true;
   }
   auto badpath = prefix + "_bad.osh";
   binary::write(badpath, mesh);
-  std::cout << "This run, stored at \"" << badpath << "\",\n";
-  std::cout << "does not match the gold at \"" << goldpath << "\"\n";
+  if (comm->rank() == 0) {
+    std::cout << "This run, stored at \"" << badpath << "\",\n";
+    std::cout << "does not match the gold at \"" << goldpath << "\"\n";
+  }
   return false;
 }
